Split crates into base crate, and impl crate (providing proc-macro)
