+++ nom-derive-impl/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "nom-derive"
+name = "nom-derive-impl"
 description = "Custom derive nom parsers from struct"
 version = "0.8.0"
 license = "MIT/Apache-2.0"
@@ -8,7 +8,7 @@ categories = ["parsing"]
 authors = ["Pierre Chifflier <chifflier@wzdftpd.net>"]
 homepage = "https://github.com/rust-bakery/nom-derive"
 repository = "https://github.com/rust-bakery/nom-derive.git"
-readme = "README.md"
+readme = "../README.md"
 edition = "2018"
 
 include = [
@@ -19,15 +19,16 @@ include = [
   "Cargo.toml",
   "src/*.rs",
   "src/meta/*.rs",
+  "tests/*.rs",
 ]
 
-[workspace]
-members = ["nom-derive-impl"]
+[lib]
+proc-macro = true
 
 [dependencies]
-nom = "6.0"
-nom-derive-impl = { version="=0.8.0", path="./nom-derive-impl" }
+syn = { version=">=1.0.58", features=["parsing","extra-traits","full"] }
+quote = "1.0"
+proc-macro2 = "1.0"
 
-[dev-dependencies]
-pretty_assertions = "0.7"
-trybuild = "1.0"
+# [dev-dependencies]
+# nom = "6.0"
